backend/requirements.txt
openai-whisper
flask
flask-cors
pydub
deep-translator
